+++ Release Notes.md
@@ -1,7 +1,7 @@
 # Release notes
 
 
-## 1.2.0
+## 1.1.1
 
 ### New views: